upload and remove images

+++ src/controllers/recipes.controller.js
@@ -1,7 +1,11 @@
 const Recipe = require('../models/Recipe')
+const path = require('path')
+const { randomNumber } = require('../helpers/libs')
+const fs = require('fs-extra')
 
 const addRecipes = async ( req, res ) => {
     const { title, description } = req.body
+
     const errors = []
     if(!title){
         errors.push({text: 'Please type a Title'})
@@ -13,15 +17,46 @@ const addRecipes = async ( req, res ) => {
         res.render('recipes/add-recipes', {
             errors,
             title,
-            description
+            description,
         })
     }
     else{
-        const newRecipe = new Recipe({title, description})
-        await newRecipe.save()
-        req.flash('success_msg', 'Recipe Added Successfully')
-        res.redirect('/recipes')
+
+        const saveImage = async () =>{
+            const imageURL = randomNumber()
+            const images = await Recipe.find({filename: imageURL})
+            if(images.length > 0){
+                saveImage()
+            }else{
+                const ext = path.extname(req.file.originalname).toLocaleLowerCase()
+                const imageTempPath = req.file.path
+                const targetPath = path.resolve(`src/public/upload/${imageURL}${ext}`)
+                if(ext === '.png' || ext === '.jpg' || ext === '.jpeg') {
+                    await fs.rename(imageTempPath, targetPath)
+                    const newRecipe = new Recipe({
+                        title, 
+                        description,
+                        filename: imageURL + ext
+                    })
+                    await newRecipe.save()
+                    req.flash('success_msg', 'Recipe Added Successfully')
+                    res.redirect('/recipes')
+                }
+                else{
+                    await fs.unlink(imageTempPath)
+                    req.flash('error_msg', 'Only Images are allowed')
+                    res.redirect('/recipes/add')
+                }
+            }
+
+        }
+
+        saveImage()
+
+        
+        
     }
+        
 }
 
 const editRecipe = async ( req, res ) => {
@@ -32,9 +67,13 @@ const editRecipe = async ( req, res ) => {
 }
 
 const deleteRecipe = async ( req, res ) => {
-    await Recipe.findByIdAndDelete(req.params.id)
-    req.flash('success_msg', 'Recipe Delete Successfully')
-    res.redirect('/recipes')
+    const recipe = await Recipe.findById(req.params.id)
+    if(recipe){
+        await fs.unlink(`src/public/upload/${recipe.filename}`)
+        await Recipe.findByIdAndDelete(req.params.id)
+        req.flash('success_msg', 'Recipe Delete Successfully')
+        res.redirect('/recipes')
+    }
 }
 
 module.exports = {
